# updater: Hash files with multiple threads