### Description

Breaks out the hashing of existing files into a separate step that runs with multiple threads.

On my system this results in about 4x faster hashing when the files are not cached (7800 ms -> 1900 ms) and still about 2x faster hashing when files are in the OS cache (980 ms -> 480 ms).

### Motivation and Context

Faster hashing. Further speeding up the updating process.

### How Has This Been Tested?

Tested locally with changes applied on top of #8074.

### Types of changes

- New feature (non-breaking change which adds functionality)

### Checklist:

- [x] My code has been run through [clang-format](https://github.com/obsproject/obs-studio/blob/master/.clang-format).
- [x] I have read the [**contributing** document](https://github.com/obsproject/obs-studio/blob/master/CONTRIBUTING.rst).
- [x] My code is not on the master branch.
- [x] The code has been tested.
- [x] All commit messages are properly formatted and commits squashed where appropriate.
- [x] I have included updates to all appropriate documentation.